chore: :truck: moves UserNotFound to correct folder
